[ISSUE-165] 백준 18870번 좌표 압축   - 백준 18870번 좌표 압축 풀이
